move sequences/projects into files
