+++ pytest_dbt_adapter/exceptions.py
@@ -1,0 +1,6 @@
+class DBTException(Exception):
+    """ custom exception for error reporting. """
+
+
+class TestProcessingException(Exception):
+    pass
