+++ pytest_dbt_adapter/projects/base.yml
@@ -1,0 +1,27 @@
+name: base
+paths:
+    data/base.csv: files.seeds.base
+    models/view.sql: files.models.base_view
+    models/table.sql: files.models.base_table
+    models/schema.yml: files.schemas.base
+facts:
+    seed:
+        length: 1
+        names:
+            - base
+    run:
+        length: 2
+        names:
+            - view
+            - table
+    catalog:
+        nodes:
+            length: 3
+        sources:
+            length: 1
+    persisted_relations:
+        - base
+        - view
+        - table
+    base:
+        rowcount: 10
