+++ pytest_dbt_adapter/projects/schema_tests.yml
@@ -1,0 +1,32 @@
+name: schema_tests
+paths:
+    data/base.csv: files.seeds.base
+    data/schema.yml: files.schemas.test_seed
+    models/view.sql: files.models.base_view
+    models/table.sql: files.models.base_table
+    models/schema.yml: files.schemas.base
+    models/schema_view.yml: files.schemas.test_view
+    models/schema_table.yml: files.schemas.test_table
+facts:
+    seed:
+        length: 1
+        names:
+            - base
+    run:
+        length: 2
+        names:
+            - view
+            - table
+    test:
+        length: 3
+    catalog:
+        nodes:
+            length: 3
+        sources:
+            length: 1
+    persisted_relations:
+        - base
+        - view
+        - table
+    base:
+        rowcount: 10
