+++ pytest_dbt_adapter/sequences/data_test.yml
@@ -1,0 +1,12 @@
+name: data_test
+project: data_tests
+sequence:
+    - type: dbt
+      cmd: test
+      check: false
+    - type: run_results
+      length: fact.test.length
+      names: fact.test.names
+      attributes:
+        passing.fail: false
+        failing.fail: true
