+++ pytest_dbt_adapter/sequences/data_test_ephemeral_models.yml
@@ -1,0 +1,21 @@
+name: data_test_ephemeral_models
+project: data_test_ephemeral_models
+sequence:
+    - type: dbt
+      cmd: seed
+    - type: run_results
+      length: fact.seed.length
+    - type: dbt
+      cmd: test
+      check: false
+    - type: run_results
+      length: fact.test.length
+      names: fact.test.names
+      attributes:
+        passing.fail: false
+        failing.fail: true
+    - type: dbt
+      cmd: run
+    - type: run_results
+      length: fact.run.length
+      names: fact.run.names
