+++ pytest_dbt_adapter/sequences/incremental.yml
@@ -1,0 +1,41 @@
+name: incremental
+project: incremental
+sequence:
+  - type: dbt
+    cmd: seed
+  - type: run_results
+    length: fact.seed.length
+  - type: dbt
+    cmd: run
+    vars:
+      seed_name: base
+  - type: relation_rows
+    name: base
+    length: fact.base.rowcount
+  - type: run_results
+    length: fact.run.length
+  - type: relations_equal
+    relations:
+      - base
+      - incremental
+  - type: dbt
+    cmd: run
+    vars:
+      seed_name: extended
+  - type: relation_rows
+    name: extended
+    length: fact.extended.rowcount
+  - type: run_results
+    length: fact.run.length
+  - type: relations_equal
+    relations:
+      - extended
+      - incremental
+  - type: dbt
+    cmd: docs generate
+  - type: catalog
+    exists: True
+    nodes:
+      length: fact.catalog.nodes.length
+    sources:
+      length: fact.catalog.sources.length
