+++ pytest_dbt_adapter/spec_file.py
@@ -11,828 +11,13 @@ from typing import Dict, Any, Iterable
 import pytest
 import yaml
 
+from .exceptions import TestProcessingException, DBTException
+from .builtin import BUILTIN_TEST_SEQUENCES, DEFAULT_PROJECTS, DbtProject
+
+
 from dbt.adapters.factory import FACTORY
 from dbt.config import RuntimeConfig
 from dbt.main import parse_args
-
-
-# custom test loader
-def pytest_collect_file(parent, path):
-    if path.ext == ".dbtspec":  # and path.basename.startswith("test"):
-        return DbtSpecFile.from_parent(parent, fspath=path)
-
-
-def pytest_addoption(parser):
-    group = parser.getgroup('dbtadapter')
-    group.addoption(
-        '--no-drop-schema',
-        action='store_false',
-        dest='drop_schema',
-
-    )
-
-
-DEFAULT_DBT_PROJECT = {
-    'name': 'dbt_test_project',
-    'config-version': 2,
-    'version': '1.0.0',
-}
-
-
-NAMES_BASE = """
-id,name,some_date
-1,Easton,1981-05-20T06:46:51
-2,Lillian,1978-09-03T18:10:33
-3,Jeremiah,1982-03-11T03:59:51
-4,Nolan,1976-05-06T20:21:35
-5,Hannah,1982-06-23T05:41:26
-6,Eleanor,1991-08-10T23:12:21
-7,Lily,1971-03-29T14:58:02
-8,Jonathan,1988-02-26T02:55:24
-9,Adrian,1994-02-09T13:14:23
-10,Nora,1976-03-01T16:51:39
-""".lstrip()
-
-
-NAMES_EXTENDED = NAMES_BASE + """
-11,Mateo,2014-09-07T17:04:27
-12,Julian,2000-02-04T11:48:30
-13,Gabriel,2001-07-10T07:32:52
-14,Isaac,2002-11-24T03:22:28
-15,Levi,2009-11-15T11:57:15
-16,Elizabeth,2005-04-09T03:50:11
-17,Grayson,2019-08-06T19:28:17
-18,Dylan,2014-03-01T11:50:41
-19,Jayden,2009-06-06T07:12:49
-20,Luke,2003-12-05T21:42:18
-""".lstrip()
-
-
-NAMES_ADD_COLUMN = """
-id,name,some_date,last_initial
-1,Easton,1981-05-20T06:46:51,A
-2,Lillian,1978-09-03T18:10:33,B
-3,Jeremiah,1982-03-11T03:59:51,C
-4,Nolan,1976-05-06T20:21:35,D
-5,Hannah,1982-06-23T05:41:26,E
-6,Eleanor,1991-08-10T23:12:21,F
-7,Lily,1971-03-29T14:58:02,G
-8,Jonathan,1988-02-26T02:55:24,H
-9,Adrian,1994-02-09T13:14:23,I
-10,Nora,1976-03-01T16:51:39,J
-""".lstrip()
-
-
-class Model:
-    def __init__(self, config, body):
-        self.config = config
-        self.body = body
-
-    @classmethod
-    def from_dict(cls, dct):
-        try:
-            config = dct.get('config', {})
-            if 'materialized' in dct:
-                config['materialized'] = dct['materialized']
-            return cls(config=config, body=dct['body'])
-        except KeyError as exc:
-            raise TestProcessingException(
-                f'Invalid test, model is missing key {exc}'
-            )
-
-    def config_params(self):
-        if not self.config:
-            return ''
-        else:
-            pairs = ', '.join(
-                '{!s}={!r}'.format(key, value)
-                for key, value in self.config.items()
-            )
-            return '{{ config(' + pairs + ') }}'
-
-    def render(self):
-        return '\n'.join([self.config_params(), self.body])
-
-
-class DbtProject:
-    def __init__(
-        self,
-        name: str,
-        dbt_project_yml: str,
-        paths: Dict[str, str],
-        facts: Dict[str, Any]
-    ):
-        self.name = name
-        self.dbt_project_yml = dbt_project_yml
-        self.paths = paths
-        self.facts = facts
-
-    def write(self, path: str):
-        project_path = os.path.join(path, 'project')
-        os.makedirs(project_path)
-        with open(os.path.join(project_path, 'dbt_project.yml'), 'w') as fp:
-            fp.write(yaml.safe_dump(self.dbt_project_yml))
-
-        for relpath, contents in self.paths.items():
-            fullpath = os.path.join(project_path, relpath)
-            os.makedirs(os.path.dirname(fullpath), exist_ok=True)
-
-            if isinstance(contents, str) and contents.startswith('files.'):
-                contents_path = contents.split('.')[1:]
-                cur = KNOWN_FILES
-                for part in contents_path:
-                    if part not in cur:
-                        raise TestProcessingException(
-                            f'at known file lookup {contents}, could not find '
-                            f'part {part} in known files path'
-                        )
-                    cur = cur[part]
-                contents = cur
-
-            if relpath.startswith('models/') and relpath.endswith('.sql'):
-                if isinstance(contents, dict):
-                    model = Model.from_dict(contents)
-                    contents = model.render()
-            if not isinstance(contents, str):
-                raise TestProcessingException(f'{contents} is not a string')
-
-            with open(fullpath, 'w') as fp:
-                fp.write(contents)
-
-    @classmethod
-    def from_dict(cls, dct, overriding=None):
-        if overriding is None:
-            overriding = {}
-
-        paths: Dict[str, Any]
-        facts: Dict[str, Any]
-        dbt_project_yml: Dict[str, Any]
-
-        if 'overrides' in dct:
-            name = dct['overrides']
-            if name not in overriding:
-                raise TestProcessingException(
-                    f'Invalid project definition, override name {name} not '
-                    'known'
-                ) from None
-            dbt_project_yml = overriding[name].dbt_project_yml.copy()
-            paths = overriding[name].paths.copy()
-            facts = overriding[name].facts.copy()
-        else:
-            try:
-                name = dct['name']
-            except KeyError:
-                raise TestProcessingException(
-                    f'Invalid project definition, no name in {dct}'
-                ) from None
-
-            dbt_project_yml = DEFAULT_DBT_PROJECT.copy()
-            paths = {}
-            facts = {}
-
-        dbt_project_yml.update(dct.get('dbt_project_yml', {}))
-
-        paths.update(dct.get('paths', {}))
-        facts.update(dct.get('facts', {}))
-        return cls(
-            name=name,
-            dbt_project_yml=dbt_project_yml,
-            paths=paths,
-            facts=facts,
-        )
-
-
-SEED_SOURCE_YML = """
-version: 2
-sources:
-  - name: raw
-    schema: "{{ target.schema }}"
-    tables:
-      - name: seed
-        identifier: "{{ var('seed_name', 'base') }}"
-"""
-
-TEST_SEEDS_SCHEMA_YML_TEST_BASE = """
-version: 2
-models:
-  - name: base
-    columns:
-     - name: id
-       tests:
-         - not_null
-"""
-
-TEST_MODELS_SCHEMA_YML_TEST_VIEW = """
-version: 2
-models:
-  - name: view
-    columns:
-     - name: id
-       tests:
-         - not_null
-"""
-
-TEST_MODELS_SCHEMA_YML_TEST_TABLE = """
-version: 2
-models:
-  - name: table
-    columns:
-     - name: id
-       tests:
-         - not_null
-"""
-
-
-TEST_PASSING_DATA_TEST = """
-select 1 as id where id = 2
-"""
-
-TEST_FAILING_DATA_TEST = """
-select 1 as id where id = 1
-"""
-
-
-TEST_EPHEMERAL_DATA_TEST_PASSING = '''
-with my_other_cool_cte as (
-    select id, name from {{ ref('ephemeral') }}
-    where id > 1000
-)
-select name, id from my_other_cool_cte
-'''
-
-
-TEST_EPHEMERAL_DATA_TEST_FAILING = '''
-with my_other_cool_cte as (
-    select id, name from {{ ref('ephemeral') }}
-    where id < 1000
-)
-select name, id from my_other_cool_cte
-'''
-
-INCREMENTAL_MODEL = """
-select * from {{ source('raw', 'seed') }}
-{% if is_incremental() %}
-where id > (select max(id) from {{ this }})
-{% endif %}
-""".strip()
-
-
-CC_SNAPSHOT_SQL = '''
-{% snapshot cc_snapshot %}
-    {{ config(
-        check_cols='all', unique_key='id', strategy='check',
-        target_database=database, target_schema=schema
-    ) }}
-    select * from {{ ref(var('seed_name', 'base')) }}
-{% endsnapshot %}
-'''.strip()
-
-
-TS_SNAPSHOT_SQL = '''
-{% snapshot ts_snapshot %}
-    {{ config(
-        strategy='timestamp',
-        unique_key='id',
-        updated_at='some_date',
-        target_database=database,
-        target_schema=schema,
-    )}}
-    select * from {{ ref(var('seed_name', 'base')) }}
-{% endsnapshot %}
-'''.strip()
-
-
-EPHEMERAL_WITH_CTE = """
-with my_cool_cte as (
-  select name, id from {{ ref('base') }}
-)
-select id, name from my_cool_cte where id is not null
-"""
-
-
-KNOWN_FILES = {
-    'seeds': {
-        'base': NAMES_BASE,
-        'newcolumns': NAMES_ADD_COLUMN,
-        'added': NAMES_EXTENDED,
-    },
-    'models': {
-        'base_table': {
-            'materialized': 'table',
-            'body': "select * from {{ source('raw', 'seed') }}",
-        },
-        'base_view': {
-            'materialized': 'view',
-            'body': "select * from {{ source('raw', 'seed') }}",
-        },
-        'ephemeral': {
-            'materialized': 'ephemeral',
-            'body': "select * from {{ source('raw', 'seed') }}",
-        },
-        'ephemeral_with_cte': {
-            'materialized': 'ephemeral',
-            'body': EPHEMERAL_WITH_CTE,
-        },
-        'ephemeral_view': {
-            'materialized': 'view',
-            'body': "select * from {{ ref('ephemeral') }}",
-        },
-        'ephemeral_table': {
-            'materialized': 'table',
-            'body': "select * from {{ ref('ephemeral') }}",
-        },
-        'incremental': {
-            'materialized': 'incremental',
-            'body': INCREMENTAL_MODEL,
-        }
-    },
-    'snapshots': {
-        'check_cols': CC_SNAPSHOT_SQL,
-        'timestamp': TS_SNAPSHOT_SQL,
-    },
-    'tests': {
-        'passing': TEST_PASSING_DATA_TEST,
-        'failing': TEST_FAILING_DATA_TEST,
-        'ephemeral': {
-            'passing': TEST_EPHEMERAL_DATA_TEST_PASSING,
-            'failing': TEST_EPHEMERAL_DATA_TEST_FAILING,
-        }
-    },
-    'schemas': {
-        'base': SEED_SOURCE_YML,
-        'test_seed': TEST_SEEDS_SCHEMA_YML_TEST_BASE,
-        'test_view': TEST_MODELS_SCHEMA_YML_TEST_VIEW,
-        'test_table': TEST_MODELS_SCHEMA_YML_TEST_TABLE,
-    },
-}
-
-
-EMPTY_PROJECT = yaml.safe_load('''
-    name: empty
-    facts:
-        seed:
-            length: 0
-        run:
-            length: 0
-        catalog:
-            nodes:
-                length: 0
-            sources:
-                length: 0
-''')
-
-
-BASE_PROJECT = yaml.safe_load('''
-    name: base
-    paths:
-        data/base.csv: files.seeds.base
-        models/view.sql: files.models.base_view
-        models/table.sql: files.models.base_table
-        models/schema.yml: files.schemas.base
-    facts:
-        seed:
-            length: 1
-            names:
-                - base
-        run:
-            length: 2
-            names:
-                - view
-                - table
-        catalog:
-            nodes:
-                length: 3
-            sources:
-                length: 1
-        persisted_relations:
-            - base
-            - view
-            - table
-        base:
-            rowcount: 10
-''')
-
-
-EPHEMERAL_PROJECT = yaml.safe_load('''
-    name: ephemeral
-    paths:
-        data/base.csv: files.seeds.base
-        models/ephemeral.sql: files.models.ephemeral
-        models/view.sql: files.models.ephemeral_view
-        models/table.sql: files.models.ephemeral_table
-        models/schema.yml: files.schemas.base
-    facts:
-        seed:
-            length: 1
-            names:
-                - base
-        run:
-            length: 2
-            names:
-                - view
-                - table
-        catalog:
-            nodes:
-                length: 3
-            sources:
-                length: 1
-        persisted_relations:
-            - base
-            - view
-            - table
-        base:
-            rowcount: 10
-
-''')
-
-INCREMENTAL_PROJECT = yaml.safe_load('''
-    name: incremental
-    paths:
-        data/base.csv: files.seeds.base
-        data/extended.csv: files.seeds.added
-        models/incremental.sql: files.models.incremental
-        models/schema.yml: files.schemas.base
-    facts:
-        seed:
-            length: 2
-            names:
-                - base
-                - extended
-        run:
-            length: 1
-            names:
-                - incremental
-        catalog:
-            nodes:
-                length: 3
-            sources:
-                length: 1
-        persisted_relations:
-            - base
-            - extended
-            - incremental
-        base:
-            rowcount: 10
-        extended:
-            rowcount: 20
-''')
-
-SNAPSHOT_PROJECT = yaml.safe_load('''
-    name: snapshot
-    paths:
-        data/base.csv: files.seeds.base
-        data/newcolumns.csv: files.seeds.newcolumns
-        data/added.csv: files.seeds.added
-        snapshots/cc_snapshot.sql: files.snapshots.check_cols
-        snapshots/ts_snapshot.sql: files.snapshots.timestamp
-    facts:
-        seed:
-            length: 3
-            names:
-                - base
-                - newcolumns
-                - added
-        snapshot:
-            length: 2
-            names:
-                - cc_snapshot
-                - ts_snapshot
-        base:
-            rowcount: 10
-        added:
-            rowcount: 20
-        newcolumns:
-            rowcount: 10
-        added_plus_ten:
-            rowcount: 30
-        added_plus_twenty:
-            rowcount: 40
-''')
-
-
-DATA_TESTS_PROJECT = yaml.safe_load('''
-    name: data_tests
-    paths:
-        test/passing.sql: files.tests.passing
-        test/failing.sql: files.tests.failing
-    facts:
-        test:
-            length: 2
-            names:
-                - passing
-                - failing
-''')
-
-DATA_TEST_EPHEMERALS_PROJECT = yaml.safe_load('''
-    name: data_test_ephemeral_models
-    paths:
-        data/base.csv: files.seeds.base
-        models/ephemeral.sql: files.models.ephemeral_with_cte
-        models/passing_model.sql: files.tests.ephemeral.passing
-        models/failing_model.sql: files.tests.ephemeral.failing
-        models/schema.yml: files.schemas.base
-        test/passing.sql: files.tests.ephemeral.passing
-        test/failing.sql: files.tests.ephemeral.failing
-    facts:
-        seed:
-            length: 1
-            names:
-                - base
-        run:
-            length: 2
-            names:
-                - passing_model
-                - failing_model
-        test:
-            length: 2
-            names:
-                - passing
-                - failing
-''')
-
-
-SCHEMA_TESTS_PROJECT = yaml.safe_load('''
-    name: schema_tests
-    paths:
-        data/base.csv: files.seeds.base
-        data/schema.yml: files.schemas.test_seed
-        models/view.sql: files.models.base_view
-        models/table.sql: files.models.base_table
-        models/schema.yml: files.schemas.base
-        models/schema_view.yml: files.schemas.test_view
-        models/schema_table.yml: files.schemas.test_table
-    facts:
-        seed:
-            length: 1
-            names:
-                - base
-        run:
-            length: 2
-            names:
-                - view
-                - table
-        test:
-            length: 3
-        catalog:
-            nodes:
-                length: 3
-            sources:
-                length: 1
-        persisted_relations:
-            - base
-            - view
-            - table
-        base:
-            rowcount: 10
-''')
-
-
-DEFAULT_PROJECTS = {
-    p['name']: p for p in [
-        EMPTY_PROJECT,
-        BASE_PROJECT,
-        EPHEMERAL_PROJECT,
-        INCREMENTAL_PROJECT,
-        SNAPSHOT_PROJECT,
-        DATA_TESTS_PROJECT,
-        SCHEMA_TESTS_PROJECT,
-        DATA_TEST_EPHEMERALS_PROJECT,
-    ]
-}
-
-
-BUILTIN_TEST_SEQUENCES = {
-    'empty': yaml.safe_load('''
-        project: empty
-        sequence:
-          - type: dbt
-            cmd: seed
-          - type: run_results
-            exists: False
-          - type: dbt
-            cmd: run
-          - type: run_results
-            exists: False
-          - type: catalog
-            exists: False
-          - type: dbt
-            cmd: docs generate
-          - type: run_results
-            exists: False
-          - type: catalog
-            exists: True
-            nodes:
-              length: fact.catalog.nodes.length
-            sources:
-              length: fact.catalog.sources.length
-        '''),
-    'base': yaml.safe_load('''
-        project: base
-        sequence:
-          - type: dbt
-            cmd: seed
-          - type: run_results
-            length: fact.seed.length
-          - type: dbt
-            cmd: run
-          - type: run_results
-            length: fact.run.length
-          - type: relation_rows
-            name: base
-            length: fact.base.rowcount
-          - type: relations_equal
-            relations: fact.persisted_relations
-          - type: dbt
-            cmd: docs generate
-          - type: catalog
-            exists: True
-            nodes:
-              length: fact.catalog.nodes.length
-            sources:
-              length: fact.catalog.sources.length
-        '''),
-    'ephemeral': yaml.safe_load('''
-        project: ephemeral
-        sequence:
-          - type: dbt
-            cmd: seed
-          - type: run_results
-            length: fact.seed.length
-          - type: dbt
-            cmd: run
-          - type: run_results
-            length: fact.run.length
-          - type: relation_rows
-            name: base
-            length: fact.base.rowcount
-          - type: relations_equal
-            relations: fact.persisted_relations
-          - type: dbt
-            cmd: docs generate
-          - type: catalog
-            exists: True
-            nodes:
-              length: fact.catalog.nodes.length
-            sources:
-              length: fact.catalog.sources.length
-        '''),
-    'incremental': yaml.safe_load('''
-        project: incremental
-        sequence:
-          - type: dbt
-            cmd: seed
-          - type: run_results
-            length: fact.seed.length
-          - type: dbt
-            cmd: run
-            vars:
-              seed_name: base
-          - type: relation_rows
-            name: base
-            length: fact.base.rowcount
-          - type: run_results
-            length: fact.run.length
-          - type: relations_equal
-            relations:
-              - base
-              - incremental
-          - type: dbt
-            cmd: run
-            vars:
-              seed_name: extended
-          - type: relation_rows
-            name: extended
-            length: fact.extended.rowcount
-          - type: run_results
-            length: fact.run.length
-          - type: relations_equal
-            relations:
-              - extended
-              - incremental
-          - type: dbt
-            cmd: docs generate
-          - type: catalog
-            exists: True
-            nodes:
-              length: fact.catalog.nodes.length
-            sources:
-              length: fact.catalog.sources.length
-    '''),
-    'snapshot': yaml.safe_load('''
-        project: snapshot
-        sequence:
-          - type: dbt
-            cmd: seed
-          - type: run_results
-            length: fact.seed.length
-          - type: dbt
-            cmd: snapshot
-          - type: relation_rows
-            name: cc_snapshot
-            length: fact.base.rowcount
-          - type: relation_rows
-            name: ts_snapshot
-            length: fact.base.rowcount
-         # point at the "added" seed so the snapshot sees 10 new rows
-          - type: dbt
-            cmd: snapshot
-            vars:
-              seed_name: added
-          - type: relation_rows
-            name: cc_snapshot
-            length: fact.added.rowcount
-          - type: relation_rows
-            name: ts_snapshot
-            length: fact.added.rowcount
-         # update some timestamps in the "added" seed so the snapshot sees 10 more new rows
-          - type: update_rows
-            name: added
-            dst_col: some_date
-            clause:
-              src_col: some_date
-              type: add_timestamp
-            where: id > 10 and id < 21
-          - type: dbt
-            cmd: snapshot
-            vars:
-              seed_name: added
-          - type: relation_rows
-            name: ts_snapshot
-            length: fact.added_plus_ten.rowcount
-          - type: relation_rows
-            name: cc_snapshot
-            length: fact.added_plus_ten.rowcount
-          - type: update_rows
-            name: added
-            dst_col: name
-            clause:
-              src_col: name
-              type: add_string
-              value: _updated
-            where: id < 11
-          - type: dbt
-            cmd: snapshot
-            vars:
-              seed_name: added
-          - type: relation_rows
-            name: ts_snapshot
-            length: fact.added_plus_ten.rowcount
-          - type: relation_rows
-            name: cc_snapshot
-            length: fact.added_plus_twenty.rowcount
-    '''),
-    'data_test': yaml.safe_load('''
-        project: data_tests
-        sequence:
-            - type: dbt
-              cmd: test
-              check: false
-            - type: run_results
-              length: fact.test.length
-              names: fact.test.names
-              attributes:
-                passing.fail: false
-                failing.fail: true
-    '''),
-    'data_test_ephemeral_models': yaml.safe_load('''
-        project: data_test_ephemeral_models
-        sequence:
-            - type: dbt
-              cmd: seed
-            - type: run_results
-              length: fact.seed.length
-            - type: dbt
-              cmd: test
-              check: false
-            - type: run_results
-              length: fact.test.length
-              names: fact.test.names
-              attributes:
-                passing.fail: false
-                failing.fail: true
-            - type: dbt
-              cmd: run
-            - type: run_results
-              length: fact.run.length
-              names: fact.run.names
-
-    '''),
-    'schema_test': yaml.safe_load('''
-        project: schema_tests
-        sequence:
-            - type: dbt
-              cmd: seed
-            - type: dbt
-              cmd: test -m base
-            - type: run_results
-              length: fact.seed.length
-            - type: dbt
-              cmd: run
-            - type: dbt
-              cmd: test
-              length: fact.test.length
-    ''')
-}
 
 
 class DbtSpecFile(pytest.File):
@@ -1444,11 +629,3 @@ class DbtItem(pytest.Item):
 
     def reportinfo(self):
         return self.fspath, 0, "usecase: {}".format(self.name)
-
-
-class DBTException(Exception):
-    """ custom exception for error reporting. """
-
-
-class TestProcessingException(Exception):
-    pass
