+++ specs/spark.dbtspec
@@ -23,6 +23,10 @@ sequences:
   test_dbt_base: base
   test_dbt_ephemeral: ephemeral
   test_dbt_incremental: incremental
+  test_dbt_snapshot_strategy_timestamp: snapshot_strategy_timestamp
+  test_dbt_snapshot_strategy_check_cols: snapshot_strategy_check_cols
   test_dbt_data_test: data_test
   test_dbt_schema_test: schema_test
-  test_dbt_ephemeral_data_tests: data_test_ephemeral_models
+  # the local cluster currently tests on spark 2.x, which does not support this
+  # if we upgrade it to 3.x, we can enable this test
+  # test_dbt_ephemeral_data_tests: data_test_ephemeral_models
